revert(site): withdraw Wax Room post and nonprofit edits
