+++ site/package.json
@@ -11,7 +11,7 @@
     "build": "astro build",
     "preview": "astro preview --port 4321",
     "check": "astro check",
-    "test:refinement": "npm run build && node --test tests/conditionsDisplayMode.test.mjs tests/responsiveNav.test.mjs tests/waxRoomEntry.test.mjs tests/contentRefinements.test.mjs",
+    "test:refinement": "npm run build -- --force && node --test tests/conditionsDisplayMode.test.mjs tests/responsiveNav.test.mjs tests/contentRollback.test.mjs tests/contentRefinements.test.mjs",
     "test:sponsors": "npm run build && node --test tests/sponsorTiers.test.mjs tests/sponsorSchemaContracts.test.mjs tests/sponsorWallBuild.test.mjs tests/sponsors-page.test.mjs"
   },
   "dependencies": {
