+++ site/src/content/pages/home.yaml
@@ -11,6 +11,7 @@ cta_coming_soon_url: https://twincitiesskiclub.org/#registration
 cta_closed_label: How to register
 cta_closed_url: https://tcsc.ski/
 mission_paragraph: >-
-  TCSC is a 501(c)(3) nonprofit where young adult skiers train together twice a
-  week, race if they want to, travel to Midwest ski weekends, volunteer, and
-  organize workouts and socials beyond practice.
+  Twin Cities Ski Club is a 501(c)(3) nonprofit dedicated to fostering a
+  supportive community for young adults (ages 21-35) by promoting a healthy
+  lifestyle through cross-country ski training sessions and educational
+  programming.
