--- site/src/content/wax_entries/stone-grind-or-hotbox.mdoc
@@ -1,33 +1,0 @@
----
-slug: Stone Grind or Hotbox? When Skis Need a Second Wind
-date: 2026-07-19
-lede: >-
-  Last winter, eleven TCSC pairs headed to the shop together. Here is how to
-  think about structure, wax conditioning, or both.
----
-
-In December 2025, one TCSC member asked a practical question: who had skis that needed a second wind? Eleven pairs joined a coordinated Finn Sisu service order.
-
-For that batch, the practical recommendation was Finn Sisu's universal grind for a Midwest/manmade mix. Hotboxing was optional for skis that had not been waxed regularly. The two services do different jobs, so that order is a useful example rather than a prescription for every ski.
-
-## What a stone grind changes
-
-A stone grind permanently cuts structure into the ski base. It is physical base work, used when the existing structure needs to be renewed or matched to the snow the ski will usually see.
-
-Shop labels vary. Fine, medium, and coarse commonly describe structure grades for different snow ranges. Universal is often a shop's broader-use pattern or menu label, not a fourth grade on the same scale. None of these labels is a slow-to-fast ranking. A finer structure suits a different snow range than a coarser one, while a universal pattern covers a broader mix. Tell the shop where and when you ski so it can help choose the appropriate family.
-
-## What a hotbox changes
-
-A hotbox uses controlled warmth to condition the base with wax over time. It does not cut a new structure and it does not replace a grind. It is a complementary waxing step that may make sense for a ski that has gone without regular waxing.
-
-## Five useful things to tell the shop
-
-1. Whether the ski is classic or skate, because technique affects structure selection, and where it is used most, including whether the snow is mainly natural, manmade, or a mix.
-2. The usual snow temperature and moisture range, not just the forecast for one outing.
-3. When the base was last ground and which structure it received, if known.
-4. How regularly the ski has been waxed and how the base has been cared for between uses.
-5. What prompted the service request, such as worn structure, base damage, or a dry-looking base.
-
-## The short version
-
-Grind when a ski's structure needs renewal. Ask about hotboxing when its wax conditioning may help. Choose both only when both jobs make sense for that ski.
