+++ site/src/pages/sponsors.astro
@@ -22,7 +22,6 @@ const mailtoHref = `mailto:${page.data.contact_email}`;
   description={metaDescription(page.data.intro)}
   headline={page.data.headline}
   subhead={page.data.intro}
-  facts={[{ value: '501(c)(3) nonprofit', label: 'Status' }]}
 >
   <div class="safe-inline-6-10 mx-auto max-w-5xl px-6 md:px-10 py-16 md:py-20">
     <SponsorWall headingLevel="h2" />
